Move Point2d/3d/4d into their own file
